==> 01-Introduction.Rmd <==
# Introduction


`r knitr::include_graphics('magiclogo.png')`


**Magic: The Gathering** is a collectible card game that has entertained millions since its inception in 1993. Over the past 27 years, there have been hundreds of sets, and tens of thousands of cards printed, each showcasing distinct artworks, attributes and flavor text. Players engage in competitive games using these cards, in a fantastical, strategic battle that tests a player’s skill, experience and deck-building ability. Subsequently, the long standing popularity of the game has driven up demand for buying, trading and selling cards through card shops and online retailers. Because of this, there are online records for many of these cards and their respective attributes. From a data standpoint, Magic: The Gathering becomes a perfect candidate for some fun, and rewarding exploratory data analysis.
