==> 04-Data-Cleaning.Rmd <==
# Data Cleaning

```{r, message=FALSE, warning=FALSE, include=FALSE}

###############################################
##########    ###  LIBRARIES  ###    ##########
###############################################

library(tidyverse)
library(plotly)
library(ggplot2)
library(gganimate)
library(magick)
library(gifski)
library(png)
library(knitr)
library(PerformanceAnalytics)
library(ggpubr)
library(lubridate)
library(ggthemes)
library(extrafont)
library(tm)
library(tidytext)
library(textdata)
library(gridExtra)
library(scales)
library(wordcloud)
library(reshape2)
library(textstem)
library(RColorBrewer)
library(echarts4r)
library(devtools)
library(rayshader)
library(kableExtra)
library(stringi)

# SETTING UP FORMATTING 

kableFormat <- function(dat, caption) {
  kable(dat, "html", escape = FALSE, caption = caption) %>%
  kable_styling(bootstrap_options = c("striped", "condensed", "bordered"),
                full_width = FALSE)
}

```


## The Cleaning Script

The script below reads in the csv's needed and, at the end, outputs a merged CSV file containing the card and 2019 prices data joined by applicable columns. The inline comments explain what the script is doing at each stage. This data will later be compared with the 2020 price data, which is fundamentally a very different data set, though it has some variables that can be directly compared. This will be expanded on during the analysis. 

```{r, message=FALSE, warning=FALSE}
###############################################
##########    ###    CARDS    ###    ##########
###############################################

# READING IN THE DATA
cardsRaw <- read_csv("cards.csv", col_names = TRUE)
# head(cardsRaw)

#SELECTING RELEVANT COLUMNS
keepCols <- c("artist","uuid", "mcmId", "mtgjsonV4Id", "multiverseId", "name", "artist", "type", "subtypes", "supertypes", "manaCost", "convertedManaCost", "keywords", "text", "flavorText",  "power", "toughness", "rarity", "edhrecRank", "isOnlineOnly")
cards <- select_(cardsRaw, .dots = keepCols)
# head(cards)

# CLEANING TYPES
cards$type <- word(cards$type, 1, sep=" —")

# REMOVING SPECIAL CHARACTERS FROM THE TEXT VARIABLES
cards$text <- str_replace_all(cards$text, "[[:punct:]]", "")
cards$flavorText <- str_replace_all(cards$flavorText, "[[:punct:]]", "")

# CLEANING THE MANA CLUSTERF**K
# If the manaCost is NA, replace it with Z so that it can be processed by str_detect. Converted back later.
cards$manaCost <- replace_na(cards$manaCost, "Z")

#Removing the brackets
cards$manaCost = str_replace_all(cards$manaCost, "[{}]", "")

#Creating a variable to hold the 'generic' mana cost
cards$genericManaCost <- NA

# Counting the generic mana cost for each row
  # If X is included in the data, that means the # of generic mana is determined by the game environment

c = 1
for (i in cards$manaCost){
  a = str_detect(i, ".*[0-9].*")
  if ( a == TRUE){
    if (substr(i, start = 1, stop = 1) == 'X'){
      cards$genericManaCost[c] <- "Determined in Game"
    } else {
    cards$genericManaCost[c] <- substr(i, start = 1, stop = 1) 
    }
  } else {
    cards$genericManaCost[c] <- "No Generic Cost"
  }
c = c + 1
}

# Cleaning manaCost to remove data pertaining to generic mana
cards$manaCost <- str_replace_all(cards$manaCost, 'X', '') %>%
  removeNumbers()

# HANDLING CARD COLOR
# Creating the color variable
cards$manaColor <- NA

# Splitting the letters so that they can be compared, these are put in a temporary variable 
cards$temp <- str_split(cards$manaCost, "")

# Identifying the mana color 
  # If the card deals with more than one type of mana, it is assigned to the category "Multiple Types"
c = 1
for (i in cards$temp){
  if (length(i) <= 1){
    cards$manaColor[c] <- i[1]
  } else if(i[1] == i[2]) {
    cards$manaColor[c] <- i[1]
  } else {
    cards$manaColor[c] <- "Multiple Types"
}
c = c+1
}

# Colored mana cost is calculated by simply taking the length of the manaCost string (i.e., WWW would be 3)
cards$colorManaCost <- nchar(cards$manaCost) 

#Handling NA's 
cards$manaColor[is.na(cards$manaColor)] <- "No Color"
cards$manaColor[cards$manaColor == 'Z'] <- "No Color"
cards$colorManaCost <- ifelse(cards$manaColor == "No Color", "No Color Cost", cards$colorManaCost)

#Removing the depreciated columns
cards <- subset(cards, select = -c(temp, manaCost))

###############################################
##########    ###    SETS     ###    ##########
###############################################

# READING IN THE DATA
setsRaw <- read_csv("sets.csv", col_names = TRUE)
# head(setsRaw)

#SELECTING RELEVANT COLUMNS
keepColsSets <- c("mcmId", "name", "releaseDate")
sets <- select_(setsRaw, .dots = keepColsSets) %>%
  rename(
    "setName" = "name"
  )
sets <- sets[!is.na(sets$mcmId), ]


###############################################
##########    ###    JOIN     ###    ##########
###############################################
# Joining cards with sets on MCM ID
mtg <- left_join(cards, sets, by = "mcmId")

# head(mtg)
# sum(is.na(mtg$releaseDate))

###############################################
##########    ##    PRICES     ##    ##########
###############################################

# READING IN THE DATA

prices19 <- read_csv("mean_2019_prices.csv", col_names = TRUE)
mtg_prices19 <- inner_join(mtg, prices19, by = c("mtgjsonV4Id" = "uuid"))

columnOrder <- c("uuid", "mtgjsonV4Id", "name", "artist", "type", "subtypes", "supertypes", "keywords", "text", "flavorText", "power", "toughness", "rarity", "edhrecRank", "isOnlineOnly", "convertedManaCost", "genericManaCost", "manaColor", "colorManaCost", "setName", "releaseDate", "mtgo", "mtgoFoil", "paper", "paperFoil")
mtg_prices19 <- mtg_prices19[, columnOrder]

# write.csv(mtg_prices19, "cleanData_New.csv")

```

### Post-Download Formatting
Because certain facets of the data cannot be retained when the data is formatted as a .CSV, we produced a 'post-download formatting' block which can be copied and pasted in at the beginning of any R file using the above CSV to convert the data to the ideal format for manipulation. 

```{r, message=FALSE, warning=FALSE}
# TOKENIZING SUBTYPES
mtg$subtypes <- str_split(mtg$subtypes, ",")

# TOKENIZING KEYWORDS
mtg$keywords <- str_split(mtg$keywords, ",")

# TURNING RARITY INTO A FACTOR 
mtg$rarity <- factor(mtg$rarity, levels = c("common", "uncommon", "rare", "mythic"), ordered = TRUE)

#FORMATTING POWER AND TOUGHNESS CORRECTLY
  # forces some to numeric, however upon investigation the cards turned to NA's are 'booster' cards, which are like spell cards
  # these cards can be identified by their key words
mtg$power <- as.numeric(mtg$power)
mtg$toughness <- as.numeric(mtg$toughness)

```

```{r}

head(mtg, 5) %>%
  kableFormat("Clean Magic the Gathering Data")

```
