==> 08-Discussion.Rmd <==
# Discussion

## Future Directions
The results of the exploratory data analysis imply that topic modeling (LDA; Latent Derelict Allocation) would be fruitful, if only just to satisfy curiosity. 

## What We Learned Along the Way

**Kevin**  
  
In terms of data analysis methods, this project has contributed substantially to my understanding of Plotly functionality, interactive plot production, and general good practices for displaying data in an accessible/readable way. Prior to generating these Magic the Gathering plots, I had little to no experience in creating interactive figures, but am now comfortable making box plots, histograms, and 3D scatter plots with “hover-text” and zooming options, and I am sure these new skills will translate well to numerous other projects and plot types.

As for what I learned from the data, I can safely tell you that Magic the Gathering has a massive amount of cards that are mostly cheap and expendable, but some that are so expensive you might want to reevaluate your life if you are considering purchasing them. There was substantial variability in the card prices, leading to some terribly low correlations between price and most other card attributes. However, I did find some interesting info about about average card price by artist (apparently April Lee and a couple other artists are producing art for some extraordinarily valuable cards). I also learned that (surprise, surprise!) mana cost correlates pretty directly with other card attributes, like power and toughness. I also confirmed my suspicion that “rare” and “mythic” cards have a tendency to be a bit pricier than “common” and “uncommon”.

**Sara**    
  
I gained an incredible amount of respect for anyone who has ever knit a Bookdown which complies code from multiple authors. Other than my near-defeat at the hands of Knitr, this was a very enjoyable project. It was less focused than my last project for this class, which turned into a small report, but that allowed us more creativity with how we explored the data and let us explore the data in more thorough detail. Rather than simply producing a single plot and crying - “Look! Demographics” - our group looked at most every aspect of the data which seemed significant. I was also able to mess around with text data, which I haven’t been able to do in a while, which was fun. I did not have the time to preform topic modeling, which I wanted to, because we were short on time, but I suppose I’ll have to wait until Machine Learning to scratch that itch. Regardless, I did a lot of munging with the data for this project and it was fun because I quite like problem solving and that’s what munging distills down to. I also enjoyed playing around with plotly and the raytracer packages - though our raytraced graphic is perhaps not the best use of the raytracer. I knew absolutely nothing about Magic the Gathering coming into this project, and am happy to have learned something about what is apparently a quite lucrative card game while investigating data on it. 

My contributions to the project were aiding in the project direction, gathering the cards data, cleaning and combining the cards and 2019 market data, completing the text analysis portion, completing the data quality portion, assisting with the write-ups, taking point on arranging group meetings, and putting together the Bookdown. This is the first time I’ve made a Bookdown alone, which I’m pretty, to be colloquial, hype about. 

**Conor**  
  
This project was so enjoyable for me, for many reasons.  I have had an interest in card games ever since the 90’s when Pokemon infiltrated the minds of every child in America.  Magic: The Gathering is a game I have enjoyed for a number of years, and getting the opportunity to explore it deeper through heaps of data with the aid of some of my hard-working peers has been a great experience.  I have not only learned a lot about the cards data, but have expanded my knowledge on data visualization techniques and had the opportunity to think critically about the most effective methods for conveying my findings to others.  Interactive plots and animated/timelapse plots are a new tool that I will surely be excited to utilize throughout my future data science academia and career. 

**Miles**  
  
I had played Magic the Gathering when I was younger and I found it fascinating to see how the card game I remembered could have such a following that it generates market data. I was not entirely surprised that a majority of the cards were in the lower price range, however, I was certainly surprised by how much someone would be willing to pay for a piece of paper or, more surprisingly, some bytes. One thing that I found personally interesting was how we could see the foil variant of the cards appearing and becoming popular. I remember witnessing that while it was happening and being able to see it in the data felt rewarding. 

This was the first opportunity that I had to unpack and work with data in JSON format. I marveled at the amount of information that can be stored in deeper and deeper levels and I gained some knowledge on manipulating that data and how to stick it together. I found that I was much more successful with this manipulation in Python instead of R but I really began to enjoy manipulating the data and creating visualization in R. It is great to be able to create informative, aesthetically pleasing, and even interactive plots and the visualization that came from my group impressed me.
